Consolidate microprofile integration tests #433
