+++ integration-tests/pom.xml
@@ -92,8 +92,7 @@
         <module>jackson</module>
         <module>jdbc</module>
         <module>mail</module>
-        <module>microprofile-health</module>
-        <module>microprofile-metrics</module>
+        <module>microprofile</module>
         <module>netty</module>
         <module>netty-http</module>
         <module>opentracing</module>
